requirements.txt
streamlit>=1.24.0
openai>=0.27.8
python-dotenv>=1.0.0
pandas>=2.0.3
SQLAlchemy>=2.0.19